test generation fixes
* improved handling of input
* script checks return value of java command
* fix symbolic link and messages

+++ scripts/generatetests.sh
@@ -9,20 +9,28 @@ java -jar "$EVOLUDO_TEST_HOME"/target/EvoLudoTest.*.jar \
 		--tests "$EVOLUDO_TEST_TEST/references" \
 		--generate "$EVOLUDO_TEST_TEST/generators" \
 		--reports "$EVOLUDO_TEST_TEST/reports" \
-		--compress
+		--compress ;
+
+# System.exit(-1) in java translates to 255
+if (( $? == 255 ))
+	then
+		exit ;
+fi
 
 nreportsafter=$((`ls | wc -l`)) ;
 
-if (( $nreportsbefore == $nreportsafter )) 
+if (( $nreportsbefore == $nreportsafter ))
 	# all tests successfully generated
 	then
-		pushd "$EVOLUDO_TEST_TEST/references" ;
+		pushd "$EVOLUDO_TEST_TEST/references" > /dev/null;
 		# find newest directory - must be the one generated for the set of tests
 		NEWSET=$(ls -td ./*/ | head -1) ;
-		echo "new '$NEWSET': all tests successfully generated. setting current reference to new set of tests." ;
+		echo "All tests for set '$NEWSET' successfully generated." ;
+		echo "Setting current reference to new set of tests." ;
 		rm ./current ;
-		ln -s "./$NEWSET" ./current ;
-		popd ;
+		ln -s "$NEWSET" current ;
+		popd > /dev/null ;
 	else
-		echo "some tests had issues. keeping current set of reference tests." ;
+		echo "Some tests had issues." ;
+		echo "Keeping current set of reference tests." ;
 fi
